+++ src/tools.ts
@@ -1,8 +1,13 @@
-import { TFile } from "obsidian";
-import { CONFLICT_OUTPUT_FILE } from "./constants";
+import { Notice, Platform, TFile } from "obsidian";
+import {
+    CONFLICT_OUTPUT_FILE,
+    DIFF_VIEW_CONFIG,
+    SPLIT_DIFF_VIEW_CONFIG,
+} from "./constants";
 import type ObsidianGit from "./main";
-import { splitRemoteBranch } from "./utils";
 import { SimpleGit } from "./gitManager/simpleGit";
+import { getNewLeaf, splitRemoteBranch } from "./utils";
+import { GeneralModal } from "./ui/modals/generalModal";
 
 export default class Tools {
     constructor(private readonly plugin: ObsidianGit) {}
@@ -75,5 +80,75 @@ export default class Tools {
                 true
             );
         }
+    }
+
+    openDiff({
+        aFile,
+        bFile,
+        aRef,
+        bRef,
+        event,
+    }: {
+        aFile: string;
+        bFile?: string;
+        aRef: string;
+        bRef?: string;
+        event?: MouseEvent;
+    }) {
+        let diffStyle = this.plugin.settings.diffStyle;
+        if (Platform.isMobileApp) {
+            diffStyle = "git_unified";
+        }
+
+        const state = {
+            aFile: aFile,
+            bFile: bFile ?? aFile,
+            aRef: aRef,
+            bRef: bRef,
+        };
+
+        if (diffStyle == "split") {
+            void getNewLeaf(this.plugin.app, event)?.setViewState({
+                type: SPLIT_DIFF_VIEW_CONFIG.type,
+                active: true,
+                state: state,
+            });
+        } else if (diffStyle == "git_unified") {
+            void getNewLeaf(this.plugin.app, event)?.setViewState({
+                type: DIFF_VIEW_CONFIG.type,
+                active: true,
+                state: state,
+            });
+        }
+    }
+
+    async runRawCommand() {
+        const gitManager = this.plugin.gitManager;
+        if (!(gitManager instanceof SimpleGit)) {
+            return;
+        }
+        const modal = new GeneralModal(this.plugin, {
+            placeholder: "push origin master",
+            allowEmpty: false,
+        });
+        const command = await modal.openAndGetResult();
+        if (command === undefined) return;
+
+        this.plugin.promiseQueue.addTask(async () => {
+            const notice = new Notice(`Running '${command}'...`, 999_999);
+
+            try {
+                const res = await gitManager.rawCommand(command);
+                if (res) {
+                    notice.setMessage(res);
+                    window.setTimeout(() => notice.hide(), 5000);
+                } else {
+                    notice.hide();
+                }
+            } catch (e) {
+                notice.hide();
+                throw e;
+            }
+        });
     }
 }