+++ src/types.ts
@@ -55,6 +55,7 @@ export interface ObsidianGitSettings {
     showFileMenu: boolean;
     authorInHistoryView: ShowAuthorInHistoryView;
     dateInHistoryView: boolean;
+    diffStyle: "git_unified" | "split";
 }
 
 /**
@@ -242,9 +243,11 @@ export interface DiffEntry {
 export interface DiffFile {
     path: string;
     vault_path: string;
+    fromPath?: string;
+    fromVaultPath?: string;
     hash: string;
     status: string;
-    binary: boolean;
+    binary?: boolean;
 }
 
 export interface WalkDifference {
@@ -278,9 +281,29 @@ export type StatusRootTreeItem = RootTreeItem<FileStatusResult>;
 export type HistoryRootTreeItem = RootTreeItem<DiffFile>;
 
 export interface DiffViewState {
-    staged: boolean;
-    file: string;
-    hash?: string;
+    /**
+     * The repo relative file path for a.
+     * For diffing a renamed file, this is the old path.
+     */
+    aFile: string;
+
+    /**
+     * The git ref to specify which state of that file should be shown.
+     * An empty string refers to the index version of a file, so you have to specifically check against undefined.
+     */
+    aRef: string;
+
+    /**
+     * The repo relative file path for b.
+     */
+    bFile: string;
+
+    /**
+     * The git ref to specify which state of that file should be shown.
+     * An empty string refers to the index version of a file, so you have to specifically check against undefined.
+     * `undefined` stands for the workign tree version.
+     */
+    bRef?: string;
 }
 
 export enum FileType {
@@ -304,6 +327,7 @@ declare module "obsidian" {
     }
     interface View {
         titleEl: HTMLElement;
+        inlineTitleEl: HTMLElement;
     }
     interface Workspace {
         on(