+++ src/ui/history/components/logFileComponent.svelte
@@ -1,8 +1,12 @@
 <script lang="ts">
     import { setIcon, TFile } from "obsidian";
-    import { DIFF_VIEW_CONFIG } from "src/constants";
     import type { DiffFile } from "src/types";
-    import { getDisplayPath, getNewLeaf, mayTriggerFileMenu } from "src/utils";
+    import {
+        fileIsBinary,
+        getDisplayPath,
+        getNewLeaf,
+        mayTriggerFileMenu,
+    } from "src/utils";
     import type HistoryView from "../historyView";
 
     export let diff: DiffFile;
@@ -17,6 +21,14 @@
         0
     );
 
+    function mainClick(event: MouseEvent) {
+        if (fileIsBinary(diff.path)) {
+            open(event);
+        } else {
+            showDiff(event);
+        }
+    }
+
     function open(event: MouseEvent) {
         const file = view.app.vault.getAbstractFileByPath(diff.vault_path);
 
@@ -28,14 +40,12 @@
     }
 
     function showDiff(event: MouseEvent) {
-        void getNewLeaf(view.app, event)?.setViewState({
-            type: DIFF_VIEW_CONFIG.type,
-            active: true,
-            state: {
-                file: diff.path,
-                staged: false,
-                hash: diff.hash,
-            },
+        view.plugin.tools.openDiff({
+            event,
+            aFile: diff.fromPath ?? diff.path,
+            aRef: `${diff.hash}^`,
+            bFile: diff.path,
+            bRef: diff.hash,
         });
     }
 </script>
@@ -44,7 +54,7 @@
 <!-- svelte-ignore a11y-no-static-element-interactions -->
 <!-- svelte-ignore a11y-no-noninteractive-element-interactions -->
 <main
-    on:click|stopPropagation={showDiff}
+    on:click|stopPropagation={mainClick}
     on:auxclick|stopPropagation={(event) => {
         if (event.button == 2)
             mayTriggerFileMenu(
@@ -54,15 +64,13 @@
                 view.leaf,
                 "git-history"
             );
-        else showDiff(event);
+        else mainClick(event);
     }}
     on:focus
     class="tree-item nav-file"
 >
     <div
         class="tree-item-self is-clickable nav-file-title"
-        class:is-active={view.plugin.lastDiffViewState?.file ==
-            diff.vault_path && view.plugin.lastDiffViewState?.hash}
         data-path={diff.vault_path}
         data-tooltip-position={side}
         aria-label={diff.vault_path}