+++ src/ui/sourceControl/components/fileComponent.svelte
@@ -1,11 +1,15 @@
 <script lang="ts">
     import { setIcon, TFile } from "obsidian";
     import { hoverPreview } from "obsidian-community-lib";
-    import { DIFF_VIEW_CONFIG } from "src/constants";
     import type { GitManager } from "src/gitManager/gitManager";
     import type { FileStatusResult } from "src/types";
     import { DiscardModal } from "src/ui/modals/discardModal";
-    import { getDisplayPath, getNewLeaf, mayTriggerFileMenu } from "src/utils";
+    import {
+        fileIsBinary,
+        getDisplayPath,
+        getNewLeaf,
+        mayTriggerFileMenu,
+    } from "src/utils";
     import type GitView from "../sourceControl";
 
     export let change: FileStatusResult;
@@ -20,6 +24,14 @@
         () => buttons.forEach((b) => setIcon(b, b.getAttr("data-icon")!)),
         0
     );
+
+    function mainClick(event: MouseEvent) {
+        if (fileIsBinary(change.path)) {
+            open(event);
+        } else {
+            showDiff(event);
+        }
+    }
 
     function hover(event: MouseEvent) {
         //Don't show previews of config- or hidden files.
@@ -48,13 +60,10 @@
     }
 
     function showDiff(event: MouseEvent) {
-        void getNewLeaf(view.app, event)?.setViewState({
-            type: DIFF_VIEW_CONFIG.type,
-            active: true,
-            state: {
-                file: change.path,
-                staged: false,
-            },
+        view.plugin.tools.openDiff({
+            aFile: change.path,
+            aRef: "",
+            event,
         });
     }
 
@@ -90,7 +99,7 @@
 <!-- svelte-ignore a11y-unknown-aria-attribute -->
 <main
     on:mouseover={hover}
-    on:click|stopPropagation={showDiff}
+    on:click|stopPropagation={mainClick}
     on:auxclick|stopPropagation={(event) => {
         if (event.button == 2)
             mayTriggerFileMenu(
@@ -100,17 +109,13 @@
                 view.leaf,
                 "git-source-control"
             );
-        else showDiff(event);
+        else mainClick(event);
     }}
     on:focus
     class="tree-item nav-file"
 >
     <div
         class="tree-item-self is-clickable nav-file-title"
-        class:is-active={view.plugin.lastDiffViewState?.file ==
-            change.vault_path &&
-            !view.plugin.lastDiffViewState?.hash &&
-            !view.plugin.lastDiffViewState?.staged}
         data-path={change.vault_path}
         data-tooltip-position={side}
         aria-label={change.vault_path}